+++ src/main/perl/t/00-credentials.t
@@ -2,10 +2,10 @@ use strict;
 use warnings;
 
 use Test::More tests => 1;
-use Data::Dumper;
 
 BEGIN {
   use_ok('Amazon::Credentials')
     or BAIL_OUT();
 }
 
+1;