+++ src/main/perl/t/04-process.t
@@ -5,18 +5,20 @@ use lib qw{ . lib};
 
 use Test::More tests => 5;
 
-use UnitTestSetup;
+use UnitTestSetup qw(:all);
+
 use Data::Dumper;
 use Cwd;
 
 BEGIN {
   use_ok('Amazon::Credentials');
-} ## end BEGIN
+}
 
 my $process = getcwd . '/get-creds-from-process';
 
-BAIL_OUT('cannot execute $process')
-  if !-x $process;
+if ( !-x $process ) {
+  BAIL_OUT("cannot execute $process");
+}
 
 init_test( test => '04-process.t', vars => { process => $process } );
 
@@ -29,11 +31,20 @@ my $creds = Amazon::Credentials->new(
 
 ok( ref $creds, 'find credentials' );
 
-like( $creds->get_aws_access_key_id, qr/^[A-Z0-9]+$/, 'aws_access_key_id' );
+like(
+  $creds->get_aws_access_key_id,
+  qr/^[[:upper:][:digit:]]+$/xsm,
+  'aws_access_key_id'
+);
 
-like( $creds->get_aws_secret_access_key,
-  qr/^[a-zA-Z0-9\+\/]+$/, 'aws_secret_access_key' )
+like(
+  $creds->get_aws_secret_access_key,
+  qr/^[[:lower:][:upper:][:digit:]+\/=]+$/xsm,
+  'aws_secret_access_key'
+) or diag( Dumper $creds);
+
+like( $creds->get_token, qr/^[[:lower:][:upper:][:digit:]\/+=]+$/xsm,
+  'token' )
   or diag( Dumper $creds);
 
-like( $creds->get_token, qr/^[a-zA-Z0-9\+\/=]+$/xsm, 'token' )
-  or diag( Dumper $creds);
+1;