+++ src/main/perl/t/08-imdsv2.t
@@ -7,10 +7,10 @@ use Test::More;
 
 if ( !$ENV{AMAZON_CREDENTIALS_TEST_ALL} ) {
   plan skip_all => 'set AMAZON_CREDENTIALS_TEST_ALL to test in AWS';
-} ## end if ( !$ENV{AMAZON_CREDENTIALS_TEST_ALL...})
+}
 else {
   plan tests => 5;
-} ## end else [ if ( !$ENV{AMAZON_CREDENTIALS_TEST_ALL...})]
+}
 
 use Data::Dumper;
 
@@ -34,3 +34,5 @@ ok(
   $new_creds->get_aws_access_key_id && $new_creds->get_aws_secret_access_key,
   'no imdsv2'
 ) or diag($new_creds);
+
+1;