+++ src/main/perl/t/10-logging.t
@@ -7,13 +7,13 @@ use Test::More;
 use Test::Output;
 
 use Data::Dumper;
+use English qw{ -no_match_vars };
 use JSON::PP;
-use UnitTestSetup;
+use UnitTestSetup qw(:all);
 
 BEGIN {
-  use English qw{ -no_match_vars };
 
-  eval <<'END_OF_TEXT';
+  eval <<'END_OF_TEXT';  ## no critic
 use Log::Log4perl;
 use Log::Log4perl::Level;
 Log::Log4perl->easy_init($DEBUG);
@@ -21,22 +21,23 @@ END_OF_TEXT
 
   if ($EVAL_ERROR) {
     plan skip_all => 'no Log::Log4perl available';
-  } ## end if ($EVAL_ERROR)
+  }
   else {
     plan tests => 2;
-  } ## end else [ if ($EVAL_ERROR) ]
+  }
 
   {
-    no strict 'refs';
+    no strict 'refs';  ## no critic
 
     *{'HTTP::Request::new'}     = sub { bless {}, 'HTTP::Request'; };
-    *{'HTTP::Request::request'} = sub { new HTTP::Response; };
+    *{'HTTP::Request::request'} = sub { HTTP::Response->new; };
 
     *{'HTTP::Response::new'}        = sub { bless {}, 'HTTP::Response'; };
     *{'HTTP::Response::is_success'} = sub { 1; };
 
     *{'LWP::UserAgent::new'}     = sub { bless {}, 'LWP::UserAgent'; };
-    *{'LWP::UserAgent::request'} = sub { new HTTP::Response; };
+    *{'LWP::UserAgent::request'} = sub { HTTP::Response->new; };
+    ## no critic
   }
 
   use Module::Loaded;
@@ -46,13 +47,15 @@ END_OF_TEXT
   mark_as_loaded(LWP::UserAgent);
 
   use_ok('Amazon::Credentials');
-} ## end BEGIN
+}
 
 init_test;
 
 my $stderr_from;
 
+########################################################################
 subtest 'logging' => sub {
+########################################################################
 
   $stderr_from = stderr_from(
     sub {
@@ -64,7 +67,7 @@ subtest 'logging' => sub {
     }
   );
 
-  ok( $stderr_from =~ /Amazon::Credentials::Logger/, 'use default logger' )
+  ok( $stderr_from =~ /Amazon::Credentials::Logger/xsm, 'use default logger' )
     or diag($stderr_from);
 
   $stderr_from = stderr_from(
@@ -78,6 +81,8 @@ subtest 'logging' => sub {
     }
   );
 
-  ok( $stderr_from =~ /using Log::Log4perl::Logger/, 'use Log::Log4perl' )
+  ok( $stderr_from =~ /using\sLog::Log4perl::Logger/xsm, 'use Log::Log4perl' )
     or diag($stderr_from);
 };
+
+1;